+++ docs/plan/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 2026-08-04
 
+- docs/plan/high-fantasy-dungeons.md
 - docs/plan/wild-oikoumene-frontier.md
 
 2026-08-03